pyproject.toml
[tool.poetry]
name = "bc-amcom"
version = "0.1.0"
description = ""
authors = ["Luis Guerreiro <luiscvlh11@gmail.com>"]
readme = "README.md"
packages = [{include = "bc_amcom"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django = "^4.2.6"
python-dotenv = "^1.0.0"
pytest-mock = "^3.11.1"
django-cors-headers = "^4.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
bandit = "^1.7.5"
pytest-django = "^4.5.2"
model-bakery = "^1.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
